+++ src/components/BaseComponent.tsx
@@ -2,7 +2,9 @@
  * @file Contains components that is the base structure.
  */
 import Box from '@mui/material/Box';
-import React from 'react';
+import Button from '@mui/material/Button';
+import ButtonGroup from '@mui/material/ButtonGroup';
+import React, { useState } from 'react';
 import { BoxPlotChart, BarChart } from './Charts';
 import InfoBox from './InfoBox';
 import { Site } from './SiteInterface';
@@ -17,6 +19,34 @@ interface InfoContainerProps {
   // sites: Array<string>;
   fileHandler: any;
 }
+
+enum ScaleTypes {
+  'Log',
+  'Linear',
+  'Percent',
+}
+
+
+
+
+/**
+ * TODO:
+ *
+ * @returns A HTML element with three MUI buttons
+ */
+function ScaleTypeButton(
+  onClickFunc: (scaleType: ScaleTypes) => void
+): JSX.Element {
+  return (
+    <ButtonGroup variant="contained" aria-label="outlined primary button group">
+      <Button onClick={() => onClickFunc(ScaleTypes.Linear)}>Linear</Button>
+      <Button onClick={() => onClickFunc(ScaleTypes.Log)}>Log</Button>
+      <Button>%</Button>
+    </ButtonGroup>
+  );
+}
+
+
 /**
  * Component that contains all Box graphs
  *
@@ -27,6 +57,12 @@ export function BoxGraphComponent(props: GraphComponentProps): JSX.Element {
   const { metrics } = props;
   const { siteProps } = props;
   const { fileHandler } = props;
+
+  const [graphScaleType, setGraphScaleTypeBox] = useState(ScaleTypes.Linear);
+
+  const handleButtonClick = (scaleType: ScaleTypes) => {
+    setGraphScaleTypeBox(scaleType);
+  };
 
   const getSiteIds = (): string[] =>
     Array.from(siteProps.keys()).filter((key) => siteProps.get(key)?.enabled);
@@ -45,6 +81,7 @@ export function BoxGraphComponent(props: GraphComponentProps): JSX.Element {
         },
       }}
     >
+      {ScaleTypeButton(handleButtonClick)}
       <Box>
         {' '}
         <BoxPlotChart
@@ -52,11 +89,44 @@ export function BoxGraphComponent(props: GraphComponentProps): JSX.Element {
           sites={getSiteIds()}
           fileHandler={fileHandler}
           siteProps={siteProps}
+          getScaleProps={getScaleProps}
         />{' '}
       </Box>
     </Box>
   );
 }
+
+/**
+ * TODO
+ * @returns
+ */
+function getScaleProps(graphScaleType) {
+  switch (graphScaleType) {
+    case ScaleTypes.Linear: {
+      return {
+        scale: { x: 'linear', y: 'linear' }, // default is already this but to make the code more readable
+        minDomain: { y: 0 },
+      };
+    }
+    case ScaleTypes.Log: {
+      return {
+        scale: { x: 'linear', y: 'log' },
+        minDomain: { y: 0.5 }, // default is y=0 but then the graph is wacky
+      };
+    }
+    case ScaleTypes.Percent: {
+      console.warn('TODO: NO PERCENT FUNCTION MADE!');
+      break;
+    }
+    default: {
+      console.warn('A type not supported was called!');
+      break;
+    }
+  }
+  // if something breaks return empty
+  return {};
+}
+
 
 /**
  * Component that contains all bar graphs
